Check if the config file is available in the start scripts.

+++ kafka-service/scripts/startKafkaDispatcherSSH.sh
@@ -2,6 +2,7 @@
 
 app_folder=$1
 
+# check if the app folder exists
 if [[ -z "$app_folder" ]]; then
       echo "ERROR: config folder not set"
       echo "Usage: startKafkaDispatcherSSH.sh <config folder>"
@@ -10,11 +11,19 @@ else
       echo "Using ${app_folder} as config folder"
 fi
 
+# check if the config file is available
+if [[ ! -e ${app_folder}/application.yml ]]; then
+    echo "ERROR: file ${app_folder}/application.yml not found."
+    exit 1
+fi
+
+# check if another instance is running
 if [[ -e .pid ]]; then
     echo "A Kafka-Dispatcher instance seems already running from this folder."
     echo "Stop the container first or remove the .pid file if the container is not running."
     exit 1
 fi
+
 mkdir -p ${app_folder}/log/
 name="kd$RANDOM"
 echo "$name" > .pid
